requirements.txt
fastapi==0.85.0
itsdangerous==2.1.2
pydantic==1.10.2
uvicorn==0.18.3
starlette~=0.20.4
python-dotenv~=0.21.0